requirements.txt
pillow
flask
flask-wtf
pyyaml
azure-storage-blob
azure.cognitiveservices.vision.face
celery